pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codemap"
version = "0.1.0"
description = "Code impact analyzer and dependency mapper for Python projects"
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "CodeMap Contributors"}]
requires-python = ">=3.11"
keywords = ["code-analysis", "dependency-graph", "impact-analysis", "python"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "click>=8.1",
    "pyan3>=1.2.0",
    "networkx>=3.0",
    "python-mermaid>=0.1.0",
    "toml>=0.10.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.1",
    "mypy>=1.7",
    "pre-commit>=3.0",
    "types-toml>=0.10.8.7",
    "types-click>=8.1",
]

[project.scripts]
codemap = "codemap.cli:cli"

[tool.setuptools]
packages = ["codemap", "codemap.analyzer", "codemap.output", "codemap.hooks"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "W", "I", "N"]

[tool.ruff.isort]
known-first-party = ["codemap"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "--strict-markers"
markers = [
    "benchmark: marks tests as performance benchmarks (deselect with '-m \"not benchmark\"')",
]
